Make Chain the base class, removing the unnecessary Corpus
