+++ lib/markovian/importers/twitter/csv_importer.rb
@@ -17,8 +17,8 @@ module Markovian
           tweet_enumerator.reject {|t| t.empty?}
         end
 
-        def corpus
-          Corpus::Compiler.new.build_corpus(texts_for_markov_analysis)
+        def chain
+          Chain::Compiler.new.build_chain(texts_for_markov_analysis)
         end
 
         protected
